requirements.txt
asgiref==3.7.2
dj-rest-auth==5.0.2
Django==4.2.6
django-allauth==0.58.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
flake8==6.0.0
sqlparse==0.4.4
psycopg2==2.9.9
psycopg2-binary>=2.9.0
Pillow==9.1.1
